+++ assignment2/.idea/misc.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ProjectRootManager" version="2" languageLevel="JDK_20" project-jdk-name="liberica-21" project-jdk-type="JavaSDK" />
 </project>
