+++ assignment2_chat/chat_server/client/src/Main.java
@@ -1,5 +1,7 @@
+import javafx.application.Application;
+
 public class Main {
   public static void main(String[] args) {
-    System.out.println("Hello world!");
+    Application.launch(MyApplication.class);
   }
 }
