--- assignment2_chat/chat_server/client/src/mediator/ServerModel.java
@@ -1,12 +1,0 @@
-package mediator;
-
-import utility.observer.javaobserver.NamedPropertyChangeSubject;
-
-public interface ServerModel extends NamedPropertyChangeSubject {
-  void connect();
-  void disconnect();
-  void send(String username, String message);
-  String getWholeConversation();
-  void createUser(String username, String password) throws IllegalArgumentException;
-
-}
