--- assignment2_chat/chat_server/client/src/model/Package.java
@@ -1,6 +1,0 @@
-package model;
-
-public abstract class Package {
-
-  public abstract String toString();
-}
