--- assignment2_chat/chat_server/server/src/Main2.java
@@ -1,7 +1,0 @@
-import javafx.application.Application;
-
-public class Main2 {
-  public static void main(String[] args) {
-    Application.launch(MyApplication.class);
-  }
-}
