+++ assignment2_chat/chat_server/server/src/MyApplication.java
@@ -3,18 +3,14 @@ import javafx.stage.Stage;
 import mediator.ChatServer;
 import model.ChatModel;
 import model.ChatModelManager;
-import view.ViewHandler;
-import viewmodel.ViewModelFactory;
+
 
 public class MyApplication extends Application {
   private ChatServer server;
-  @Override public void start(Stage primaryStage) throws Exception {
+  @Override public void start(Stage primaryStage)
+  {
     ChatModel model = new ChatModelManager();
     server=new ChatServer(model);
     new Thread(server).start();
-    //ViewModelFactory viewModelFactory = new ViewModelFactory(model);
-    //ViewHandler view = new ViewHandler(viewModelFactory);
-
-    //view.start(primaryStage);
   }
 }
