--- assignment2_chat/chat_server/server/src/observer/PropertyChangeListener.java
@@ -1,8 +1,0 @@
-package observer;
-
-import java.beans.PropertyChangeEvent;
-
-public interface PropertyChangeListener
-{
-  void propertyChange(PropertyChangeEvent evt);
-}
