--- assignment2_chat/chat_server/server/src/observer/UnnamedPropertyChangeSubject.java
@@ -1,9 +1,0 @@
-package observer;
-
-import java.beans.PropertyChangeListener;
-
-public interface UnnamedPropertyChangeSubject
-{
-  void addListener(PropertyChangeListener listener);
-  void removeListener(PropertyChangeListener listener);
-}
