--- assignment2_chat/chat_server/server/src/view/chatView.fxml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="view.ChatViewController">
-   <children>
-      <Label fx:id="headerLabel" alignment="CENTER" layoutY="14.0" prefHeight="34.0" prefWidth="600.0" text="Header" AnchorPane.topAnchor="14.0">
-         <font>
-            <Font size="22.0" />
-         </font>
-      </Label>
-      <Label fx:id="errorLabel" alignment="CENTER" layoutY="358.0" prefHeight="34.0" prefWidth="600.0" text="ErrorLabel" textFill="RED">
-         <font>
-            <Font size="16.0" />
-         </font>
-      </Label>
-      <TextArea fx:id="chatList" wrapText="true" editable="false" layoutY="52.0" prefHeight="258.0" prefWidth="600.0">
-         <font>
-            <Font size="16.0" />
-         </font></TextArea>
-      <TextField fx:id="inputField" onAction="#onEnter" layoutX="6.0" layoutY="318.0" prefHeight="25.0" prefWidth="459.0" />
-      <Button layoutX="474.0" layoutY="318.0" mnemonicParsing="false" onAction="#sendPressed" prefHeight="25.0" prefWidth="121.0" text="Send" />
-   </children>
-</AnchorPane>
