--- assignment2_chat/chat_server/server/src/viewmodel/ViewModelState.java
@@ -1,19 +1,0 @@
-package viewmodel;
-
-public class ViewModelState
-{
-  private String username;
-
-  public ViewModelState(){
-    username = "";
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
-  }
-
-  public String getUsername()
-  {
-    return username;
-  }
-}
